# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统的API文档",
        "title": "后台管理系统 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/material": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建新的材料信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "创建材料",
                "parameters": [
                    {
                        "description": "材料信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.CreateMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/application.MaterialResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/material/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据材料ID获取材料详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "获取材料详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "材料ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/application.MaterialResponse"
                        }
                    },
                    "404": {
                        "description": "材料不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据材料ID更新材料信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "更新材料信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "材料ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的材料信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.UpdateMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据材料ID删除材料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "删除材料",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "材料ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/permission/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "列出系统中所有可分配的权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "列出所有可用权限",
                "responses": {
                    "200": {
                        "description": "权限列表",
                        "schema": {
                            "$ref": "#/definitions/application.PermissionsListResponse"
                        }
                    }
                }
            }
        },
        "/permission/list-by-domain": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "列出系统中所有权限，按业务域分组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "按域分组列出权限",
                "responses": {
                    "200": {
                        "description": "权限列表",
                        "schema": {
                            "$ref": "#/definitions/application.PermissionsByDomainResponse"
                        }
                    }
                }
            }
        },
        "/permission/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取指定用户的所有个性化权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "获取用户权限列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 LoginID",
                        "name": "loginId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "权限列表",
                        "schema": {
                            "$ref": "#/definitions/application.UserPermissionsResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/permission/user/add": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "给指定用户添加个性化权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "给用户添加权限",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.AddUserPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/permission/user/remove": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定用户的个性化权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "删除用户权限",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.RemoveUserPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "application.AddUserPermissionRequest": {
            "type": "object",
            "required": [
                "loginId",
                "permission"
            ],
            "properties": {
                "loginId": {
                    "type": "string"
                },
                "permission": {
                    "type": "string"
                }
            }
        },
...
        "application.CreateMaterialRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "category": {
                    "description": "分类",
                    "type": "string",
                    "maxLength": 50
                },
                "code": {
                    "description": "原料编号（可选，不填自动生成）",
                    "type": "string",
                    "maxLength": 50
                },
                "currentPrice": {
                    "description": "当前价格",
                    "type": "number",
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "spec": {
                    "type": "string",
                    "maxLength": 200
                },
                "unit": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "application.MaterialResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentPrice": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "spec": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "application.PermissionItem": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "create",
                    "type": "string"
                },
                "desc": {
                    "description": "创建用户",
                    "type": "string"
                },
                "domain": {
                    "description": "user",
                    "type": "string"
                },
                "name": {
                    "description": "user.create",
                    "type": "string"
                }
            }
        },
        "application.PermissionsByDomainResponse": {
            "type": "object",
            "properties": {
                "domains": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/application.PermissionItem"
                        }
                    }
                }
            }
        },
        "application.PermissionsListResponse": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/application.PermissionItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "application.RemoveUserPermissionRequest": {
            "type": "object",
            "required": [
                "loginId",
                "permission"
            ],
            "properties": {
                "loginId": {
                    "type": "string"
                },
                "permission": {
                    "type": "string"
                }
            }
        },
...
        "application.UpdateMaterialRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "maxLength": 50
                },
                "code": {
                    "type": "string",
                    "maxLength": 50
                },
                "currentPrice": {
                    "type": "number",
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "spec": {
                    "type": "string",
                    "maxLength": 200
                },
                "unit": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "application.UserPermissionsResponse": {
            "type": "object",
            "properties": {
                "loginId": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "请输入 Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}